requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6

gunicorn==21.2.0
gevent==24.2.1
gevent-websocket==0.10.1

python-socketio==5.11.1
python-engineio==4.9.0

APScheduler==3.10.4

Authlib==1.3.0
requests==2.31.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.121.0
bleach==6.1.0
